requirements.txt
bs4~=0.0.1
python-dotenv~=0.20.0
selenium~=4.3.0

requests~=2.28.1
beautifulsoup4~=4.11.1